requirements.txt
pandas
numpy
pdfplumber

groq
streamlit

python-dotenv
ipykernel